backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
Pillow==10.1.0
python-decouple==3.8
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
django-storages==1.14.2
boto3==1.34.0
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
django-health-check==3.17.0
django-ratelimit==4.1.0
sentry-sdk[django]==1.39.2
setuptools==69.0.0
hiredis==2.3.2 